Make interface flags unsigned int, not unsigned long, so they aren't 64-bit even on 64-bit arches.

+++ sys/dev/ic/rtw.c
@@ -3025,7 +3025,7 @@ rtw_txring_choose(struct rtw_softc *sc, struct rtw_txsoft_blk **tsbp,
 static inline struct mbuf *
 rtw_80211_dequeue(struct rtw_softc *sc, struct ifqueue *ifq, int pri,
     struct rtw_txsoft_blk **tsbp, struct rtw_txdesc_blk **tdbp,
-    struct ieee80211_node **nip, unsigned long int *if_flagsp)
+    struct ieee80211_node **nip, unsigned int *if_flagsp)
 {
 	struct mbuf *m;
 
@@ -3057,7 +3057,7 @@ rtw_dequeue(struct ifnet *ifp, struct rtw_txsoft_blk **tsbp,
 	struct ether_header *eh;
 	struct mbuf *m0;
 	struct rtw_softc *sc;
-	unsigned long int *if_flagsp;
+	unsigned int *if_flagsp;
 
 	*mp = NULL;
